Issue 4299 - UI - fix minor issues with ldap editor
Description:  Improved how treeview handles loading subtrees with large
number of entries.  Previously, the parent entry would not be
displayed while loading its child entries, and if a timeout occurred
then the paretn entry would not be loading in the UI, and you could not
do or see anything with it.

Also added a pop modal when an error occurs when searching the database,
or the size limit is exceeded.

relates: https://github.com/389ds/389-ds-base/issues/4299

Reviewed by: spichugi(Thanks!)